:zap: Reduce web shell startup preloads
