+++ web/src/shared/ui/layout/app-loading-screen.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/shared/ui/class-name";
 
 interface AppLoadingStateProps {
   class_name?: string;
